requirements.txt
mysql-connector-python==8.4.0
sortedcontainers==2.4.0